Create PKGBUILD
This it the PKGBUILD I use on my system, it should work with all Linux distro that uses pacman (Arch Linux, Manjaro, etc.)

It install the script in `/usr/bin/find-engine` so just `find-engine path/to/game/exe` can be used, but it requires #10 to work. It currently doesn't change username and there's no maintainer filled to be used on [AUR](https://aur.archlinux.org/). I don't know if there's a license, so it's currently filled as *unknown*.